Simply ach_populate_progress param of dpa_has_achievements
Now only accepts a bool (true by default) or an int. If true and on an
author archive page, the logged in user's ID is used; otherwise the
logged in user's ID is used. If anything else is pass, it's treated as
an integer and used for the user ID