Add Thumbnail Image for Achievement Admin Page
Added filter and action to insert achievement thumbnail on the achievements admin page. Customizations includes thumbnail size, column position, and header title.

Image width or height is constrained to no more than 39px so as to not alter the height of item rows. However, should this size prove to be to small, it can easily be adjusted.