Removed Extraneous Code
I had some additional functionality in there that ended up being unnecessary and overcomplicated. I removed that line, and instead just passed the ordering information.
Before I was passing in all the args passed into this function, but cleaning them up first, so as to get all information passed through, but that is probably unnecessary at the moment.