Grunt LESS: save CSS files to src, not build.
src is a development copy of the plugin, and as such it requires
compiled CSS to load the stylesheets in the plugin during development.

When `grunt build-prod` is run, the copy:all command will copy the
generated CSS and place it in the build folder. CSSJanus, CSSMin are
then run on the copy of the CSS in the build folder.