Heartbeat: require wp-util.js
This gives helpful utility methods such as `wp.template` which we'll use to render an underscores template.
We could copy/paste this method, but wp-utils.js is a tiny file plus there may be benefits to making sure
that Achievements always uses the same underscores rendering method as WordPress itself does (for things
like the media gallery changes).