Move old notifications files into 3.5.x theme pack
The new 3.6/default won’t support backwards compatibility for the old,
no heartbeat, notifications that were deprecated in 3.5.1. Users will
have to switch to the 3.0-3.5.1 theme pack if they want to continue to
use them.

This commit re-implements this backwards compatibility change to remove
the junk out of the core files.